+++ tests/test_server.rs
@@ -806,14 +806,14 @@ async fn test_server_cookies() {
         }))
     });
 
+    let req = srv.get("/");
+    let res = req.send().await.unwrap();
+    assert!(res.status().is_success());
+
     let first_cookie = http::CookieBuilder::new("first", "first_value")
         .http_only(true)
         .finish();
     let second_cookie = http::Cookie::new("second", "second_value");
-
-    let req = srv.get("/");
-    let res = req.send().await.unwrap();
-    assert!(res.status().is_success());
 
     let cookies = res.cookies().expect("To have cookies");
     assert_eq!(cookies.len(), 2);